+++ delta-coverage-gradle/src/test/kotlin/io/github/surpsg/deltacoverage/gradle/DeltaCoveragePluginTest.kt
@@ -4,7 +4,7 @@ import io.github.surpsg.deltacoverage.gradle.unittest.applyDeltaCoveragePlugin
 import io.github.surpsg.deltacoverage.gradle.unittest.applyPlugin
 import io.github.surpsg.deltacoverage.gradle.unittest.newProject
 import io.github.surpsg.deltacoverage.gradle.unittest.testJavaProject
-import io.kotest.matchers.collections.shouldContainExactly
+import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
 import org.gradle.api.internal.project.ProjectInternal
 import org.gradle.api.plugins.JavaPlugin
 import org.junit.jupiter.api.Test
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test
 class DeltaCoveragePluginTest {
 
     @Test
-    fun `apply plugin should automatically create views from test tasks`() {
+    fun `apply plugin should automatically create aggregated view and views from test tasks`() {
         // GIVEN
         val parentProj: ProjectInternal = testJavaProject(attachSettings = true) {
             newProject {
@@ -28,6 +28,6 @@ class DeltaCoveragePluginTest {
 
         // THEN
         val config = parentProj.extensions.getByType(DeltaCoverageConfiguration::class.java)
-        config.reportViews.names.shouldContainExactly("test")
+        config.reportViews.names.shouldContainExactlyInAnyOrder("test", "aggregated")
     }
 }